Default serialization of teams is by index, Tournament class stores true teams.
