+++ src/dk/aau/cs/ds306e18/tournament/model/format/SingleEliminationFormat.java
@@ -1,8 +1,6 @@
 package dk.aau.cs.ds306e18.tournament.model.format;
 
-import com.google.gson.annotations.JsonAdapter;
 import dk.aau.cs.ds306e18.tournament.model.Team;
-import dk.aau.cs.ds306e18.tournament.model.TeamIdListAdapter;
 import dk.aau.cs.ds306e18.tournament.model.match.Match;
 import dk.aau.cs.ds306e18.tournament.model.match.MatchPlayedListener;
 import dk.aau.cs.ds306e18.tournament.model.match.MatchStatus;
@@ -11,14 +9,12 @@ import dk.aau.cs.ds306e18.tournament.ui.BracketOverviewTabController;
 import dk.aau.cs.ds306e18.tournament.ui.bracketObjects.SingleEliminationNode;
 import javafx.scene.Node;
 
-import java.lang.reflect.Array;
 import java.util.*;
 import java.util.stream.Collectors;
 
 public class SingleEliminationFormat implements Format, MatchPlayedListener {
 
     private StageStatus status = StageStatus.PENDING;
-    @JsonAdapter(TeamIdListAdapter.class)
     private ArrayList<Team> seededTeams;
     private Match finalMatch;
     private Match[] bracket;
