+++ src/dk/aau/cs/ds306e18/tournament/model/format/SwissFormat.java
@@ -1,16 +1,14 @@
 package dk.aau.cs.ds306e18.tournament.model.format;
 
-import com.google.gson.annotations.JsonAdapter;
 import dk.aau.cs.ds306e18.tournament.model.GroupFormat;
 import dk.aau.cs.ds306e18.tournament.model.Team;
-import dk.aau.cs.ds306e18.tournament.model.TeamIdToPointMapAdapter;
 import dk.aau.cs.ds306e18.tournament.model.match.Match;
 import dk.aau.cs.ds306e18.tournament.model.match.MatchChangeListener;
 import dk.aau.cs.ds306e18.tournament.model.match.MatchPlayedListener;
 import dk.aau.cs.ds306e18.tournament.model.match.MatchStatus;
-import dk.aau.cs.ds306e18.tournament.ui.bracketObjects.SwissSettingsNode;
 import dk.aau.cs.ds306e18.tournament.ui.BracketOverviewTabController;
 import dk.aau.cs.ds306e18.tournament.ui.bracketObjects.SwissNode;
+import dk.aau.cs.ds306e18.tournament.ui.bracketObjects.SwissSettingsNode;
 import javafx.scene.Node;
 
 import java.util.*;
@@ -20,7 +18,6 @@ public class SwissFormat extends GroupFormat implements MatchChangeListener, Mat
     private ArrayList<ArrayList<Match>> rounds = new ArrayList<>();
     private int maxRoundsPossible;
     private int roundCount = 4;
-    @JsonAdapter(TeamIdToPointMapAdapter.class)
     private HashMap<Team, Integer> teamPoints;
 
     transient private List<StageStatusChangeListener> statusChangeListeners = new LinkedList<>();
