+++ src/dk/aau/cs/ds306e18/tournament/serialization/TiebreakerAdapter.java
@@ -1,11 +1,11 @@
-package dk.aau.cs.ds306e18.tournament.utility;
+package dk.aau.cs.ds306e18.tournament.serialization;
 
 import com.google.gson.*;
 import dk.aau.cs.ds306e18.tournament.model.tiebreaker.TieBreaker;
 
 import java.lang.reflect.Type;
 
-public class TiebreakerAdaptor implements JsonSerializer<TieBreaker>, JsonDeserializer<TieBreaker> {
+public class TiebreakerAdapter implements JsonSerializer<TieBreaker>, JsonDeserializer<TieBreaker> {
 
     @Override
     public JsonElement serialize(TieBreaker src, Type typeOfSrc, JsonSerializationContext context) {
