+++ src/dk/aau/cs/ds306e18/tournament/utility/FileOperations.java
@@ -8,8 +8,8 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 
-import static dk.aau.cs.ds306e18.tournament.utility.Serializer.deserialize;
-import static dk.aau.cs.ds306e18.tournament.utility.Serializer.serialize;
+import static dk.aau.cs.ds306e18.tournament.serialization.Serializer.deserialize;
+import static dk.aau.cs.ds306e18.tournament.serialization.Serializer.serialize;
 
 public class FileOperations {
